Adding an e2e test for the alternative appender
